+++ server/src-lib/Hasura/Backends/Postgres/Translate/BoolExp.hs
@@ -66,13 +66,13 @@ annColExp
 annColExp rhsParser rootTable colInfoMap (ColExp fieldName colVal) = do
   colInfo <- askFieldInfo colInfoMap fieldName
   case colInfo of
-    FIColumn pgi -> AVCol pgi <$> parseBoolExpOperations rhsParser rootTable colInfoMap pgi colVal
+    FIColumn pgi -> AVColumn pgi <$> parseBoolExpOperations rhsParser rootTable colInfoMap pgi colVal
     FIRelationship relInfo -> do
       relBoolExp      <- decodeValue colVal
       relFieldInfoMap <- askFieldInfoMapSource $ riRTable relInfo
       annRelBoolExp   <- annBoolExp rhsParser rootTable relFieldInfoMap $
                          unBoolExp relBoolExp
-      return $ AVRel relInfo annRelBoolExp
+      return $ AVRelationship relInfo annRelBoolExp
     FIComputedField _ ->
       throw400 UnexpectedPayload "Computed columns can not be part of the where clause"
     -- TODO Rakesh (from master)
@@ -104,12 +104,12 @@ convColRhs
   -> AnnBoolExpFldSQL ('Postgres pgKind)
   -> State Word64 S.BoolExp
 convColRhs tableQual = \case
-  AVCol colInfo opExps -> do
+  AVColumn colInfo opExps -> do
     let colFld = fromCol @('Postgres pgKind) $ pgiColumn colInfo
-        bExps = map (mkFieldCompExp tableQual colFld) opExps
+        bExps = map (mkFieldCompExp tableQual $ LColumn colFld) opExps
     return $ foldr (S.BEBin S.AndOp) (S.BELit True) bExps
 
-  AVRel (RelInfo _ _ colMapping relTN _ _ _) nesAnn -> do
+  AVRelationship (RelInfo _ _ colMapping relTN _ _ _) nesAnn -> do
     -- Convert the where clause on the relationship
     curVarNum <- get
     put $ curVarNum + 1
@@ -124,6 +124,22 @@ convColRhs tableQual = \case
             (mkQCol tableQual lCol)
         innerBoolExp = S.BEBin S.AndOp backCompExp annRelBoolExp
     return $ S.mkExists (S.FISimple relTN $ Just $ S.Alias newIdentifier) innerBoolExp
+  AVComputedField (AnnComputedFieldBoolExp _ _ function sessionArgPresence cfBoolExp) -> do
+    case cfBoolExp of
+      CFBEScalar opExps -> do
+        -- Convert the where clause on scalar computed field
+        let bExps = map (mkFieldCompExp tableQual $ LComputedField function sessionArgPresence) opExps
+        pure $ foldr (S.BEBin S.AndOp) (S.BELit True) bExps
+      CFBETable _ boolExp -> do
+        -- Convert the where clause on table computed field
+        curVarNum <- get
+        put $ curVarNum + 1
+        let newIdentifier  = Identifier $ "_be_" <> tshow curVarNum <> "_"
+                       <> snakeCaseQualifiedObject function
+            newIdenQ = S.QualifiedIdentifier newIdentifier Nothing
+            functionExp = mkComputedFieldFunctionExp tableQual function sessionArgPresence
+                          $ Just $ S.Alias newIdentifier
+        S.mkExists (S.FIFunc functionExp) <$> convBoolRhs' newIdenQ boolExp
   where
     mkQCol q = S.SEQIdentifier . S.QIdentifier q . toIdentifier
 
@@ -155,13 +171,39 @@ foldBoolExp f = \case
   BoolExists existsExp -> foldExists existsExp
   BoolFld ce           -> f ce
 
+data LHSField b
+  = LColumn !FieldName
+  | LComputedField !QualifiedFunction !(SessionArgumentPresence (SQLExpression b))
+
+mkComputedFieldFunctionExp
+  :: S.Qual
+  -> QualifiedFunction
+  -> SessionArgumentPresence (SQLExpression ('Postgres pgKind))
+  -> Maybe S.Alias
+  -> S.FunctionExp
+mkComputedFieldFunctionExp qual function sessionArgPresence alias =
+  -- "function_schema"."function_name"("qual".*)
+  let tableRowInput = S.SEStar $ Just qual
+      functionArgs = flip S.FunctionArgs mempty $ case sessionArgPresence of
+        SAPNotPresent     -> [tableRowInput] -- No session argument
+        SAPFirst sessArg  -> [sessArg, tableRowInput]
+        SAPSecond sessArg -> [tableRowInput, sessArg]
+  in S.FunctionExp function functionArgs $ flip S.FunctionAlias Nothing <$> alias
+
 mkFieldCompExp
-  :: S.Qual -> FieldName -> OpExpG ('Postgres pgKind) S.SQLExp -> S.BoolExp
-mkFieldCompExp qual lhsField = mkCompExp (mkQField lhsField)
+  :: S.Qual -> LHSField ('Postgres pgKind) -> OpExpG ('Postgres pgKind) S.SQLExp -> S.BoolExp
+mkFieldCompExp qual lhsField = mkCompExp qLhsField
   where
+    qLhsField = case lhsField of
+      LColumn fieldName ->
+        -- "qual"."column" =
+        S.SEQIdentifier $ S.QIdentifier qual $ Identifier $ getFieldNameTxt fieldName
+      LComputedField function sessionArgPresence ->
+        -- "function_schema"."function_name"("qual".*) =
+        S.SEFunction $ mkComputedFieldFunctionExp qual function sessionArgPresence Nothing
+
     mkQCol (col, Nothing)    = S.SEQIdentifier $ S.QIdentifier qual $ toIdentifier col
     mkQCol (col, Just table) = S.SEQIdentifier $ S.mkQIdentifierTable table $ toIdentifier col
-    mkQField = S.SEQIdentifier . S.QIdentifier qual . Identifier . getFieldNameTxt
 
     mkCompExp :: SQLExpression ('Postgres pgKind) -> OpExpG ('Postgres pgKind) (SQLExpression ('Postgres pgKind)) -> S.BoolExp
     mkCompExp lhs = \case