+++ server/src-lib/Hasura/GraphQL/Schema/BoolExp.hs
@@ -71,7 +71,7 @@ boolExp sourceName tableInfo selectPermissions = memoizeOn 'boolExp (sourceName,
       P.fieldOptional fieldName Nothing <$> case fieldInfo of
         -- field_name: field_type_comparison_exp
         FIColumn columnInfo ->
-          lift $ fmap (AVCol columnInfo) <$> comparisonExps @b (pgiType columnInfo)
+          lift $ fmap (AVColumn columnInfo) <$> comparisonExps @b (pgiType columnInfo)
 
         -- field_name: field_type_bool_exp
         FIRelationship relationshipInfo -> do
@@ -80,10 +80,29 @@ boolExp sourceName tableInfo selectPermissions = memoizeOn 'boolExp (sourceName,
           let remoteTableFilter = fmapAnnBoolExp partialSQLExpToUnpreparedValue $
                                   maybe annBoolExpTrue spiFilter remotePermissions
           remoteBoolExp <- lift $ boolExp sourceName remoteTableInfo remotePermissions
-          pure $ fmap (AVRel relationshipInfo . andAnnBoolExps remoteTableFilter) remoteBoolExp
+          pure $ fmap (AVRelationship relationshipInfo . andAnnBoolExps remoteTableFilter) remoteBoolExp
 
-        -- Using computed fields in boolean expressions is not currently supported.
-        FIComputedField _ -> empty
+        FIComputedField ComputedFieldInfo{..} -> do
+          let ComputedFieldFunction{..} = _cfiFunction
+          -- For a computed field to qualify in boolean expression it shouldn't have any input arguments
+          case toList _cffInputArgs of
+            [] -> do
+              let sessionArgPresence = case _cffSessionArgument of
+                    Nothing -> SAPNotPresent
+                    Just _  -> case _cffTableArgument of
+                      FTAFirst     -> SAPSecond P.UVSession -- If table argument is first, then session argument will be second
+                      FTANamed _ 0 -> SAPSecond P.UVSession -- Index 0 => table argument is first
+                      FTANamed{}   -> SAPFirst P.UVSession -- If table argument is second, then session argument will be firest
+
+              fmap (AVComputedField . AnnComputedFieldBoolExp _cfiXComputedFieldInfo _cfiName _cffName sessionArgPresence)
+                <$> case _cfiReturnType of
+                      CFRScalar scalarType -> lift $ fmap CFBEScalar <$> comparisonExps @b (ColumnScalar scalarType)
+                      CFRSetofTable table -> do
+                        info <- askTableInfo sourceName table
+                        permissions <- lift $ tableSelectPermissions info
+                        lift $ fmap (CFBETable table) <$> boolExp sourceName info permissions
+
+            _  -> hoistMaybe Nothing
 
         -- Using remote relationship fields in boolean expressions is not supported.
         FIRemoteRelationship _ -> empty